Make result counts scroll with the panels
